passe en (int) en centimes toutes les variables de prix, de tva, de discount...  / enleve 'admin' de la selection de la userlist pour les devis  / paths des fichiers pdf passer en variables env()  / ajoute la suppression des fichiers pdf de demo en resetDemo  / ajoute lien resetDemo pour admin  / ajoute forcage de creation du fichier pdf de devis apres signature

+++ database/migrations/2015_11_15_145631_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('description');
             $table->enum('type',['time','one_shot']);
             $table->smallInteger('value')->unsigned()->default(null)->nullable();
-            $table->decimal('price')->unsigned();
+            $table->integer('price')->unsigned();
         });
     }
